Move compression in its own directory (#3586)
* move compression in its own directory

* adjust header guard for compression.h
