+++ cvmfs/cache.h
@@ -13,7 +13,7 @@
 
 #include <string>
 
-#include "compression.h"
+#include "compression/compression.h"
 #include "crypto/hash.h"
 #include "manifest.h"
 #include "util/pointer.h"
