+++ cvmfs/catalog_mgr_ro.cc
@@ -5,7 +5,7 @@
 #include "cvmfs_config.h"
 #include "catalog_mgr_ro.h"
 
-#include "compression.h"
+#include "compression/compression.h"
 #include "network/download.h"
 #include "util/exception.h"
 #include "util/posix.h"
