+++ cvmfs/catalog_sql.h
@@ -24,7 +24,7 @@
 
 #include <string>
 
-#include "compression.h"
+#include "compression/compression.h"
 #include "crypto/hash.h"
 #include "directory_entry.h"
 #include "file_chunk.h"
