+++ cvmfs/catalog_traversal.h
@@ -13,7 +13,7 @@
 #include <vector>
 
 #include "catalog.h"
-#include "compression.h"
+#include "compression/compression.h"
 #include "crypto/signature.h"
 #include "history_sqlite.h"
 #include "manifest.h"
