+++ cvmfs/compression/compression.cc
@@ -9,7 +9,7 @@
  */
 
 #include "cvmfs_config.h"
-#include "compression.h"
+#include "compression/compression.h"
 
 #include <alloca.h>
 #include <stdlib.h>
