+++ cvmfs/cvmfs_fsck.cc
@@ -23,7 +23,7 @@
 #include <cstring>
 #include <string>
 
-#include "compression.h"
+#include "compression/compression.h"
 #include "crypto/hash.h"
 #include "util/atomic.h"
 #include "util/concurrency.h"
