+++ cvmfs/file_chunk.h
@@ -17,7 +17,7 @@
 #include <vector>
 
 #include "bigvector.h"
-#include "compression.h"
+#include "compression/compression.h"
 #include "crypto/hash.h"
 #include "shortstring.h"
 #include "smallhash.h"
