+++ cvmfs/swissknife_graft.h
@@ -10,7 +10,7 @@
 #include <string>
 #include <vector>
 
-#include "compression.h"
+#include "compression/compression.h"
 #include "crypto/hash.h"
 #include "swissknife.h"
 #include "util/string.h"
