+++ cvmfs/swissknife_sync.h
@@ -8,7 +8,7 @@
 #include <string>
 #include <vector>
 
-#include "compression.h"
+#include "compression/compression.h"
 #include "repository_tag.h"
 #include "swissknife.h"
 #include "upload.h"
