+++ cvmfs/sync_mediator.h
@@ -33,7 +33,7 @@
 #include <vector>
 
 #include "catalog_mgr_rw.h"
-#include "compression.h"
+#include "compression/compression.h"
 #include "file_chunk.h"
 #include "publish/repository.h"
 #include "statistics.h"
