+++ cvmfs/upload_s3.cc
@@ -12,7 +12,7 @@
 #include <string>
 #include <vector>
 
-#include "compression.h"
+#include "compression/compression.h"
 #include "network/s3fanout.h"
 #include "options.h"
 #include "util/exception.h"
