+++ test/micro-benchmarks/CMakeLists.txt
@@ -29,7 +29,7 @@ set (CVMFS_UBENCHMARKS_SOURCES
 
   # dependencies
   ${CVMFS_SOURCE_DIR}/cache_transport.cc
-  ${CVMFS_SOURCE_DIR}/compression.cc
+  ${CVMFS_SOURCE_DIR}/compression/compression.cc
   ${CVMFS_SOURCE_DIR}/crypto/hash.cc
   ${CVMFS_SOURCE_DIR}/directory_entry.cc
   ${CVMFS_SOURCE_DIR}/glue_buffer.cc
