+++ test/unittests/t_ingestion.cc
@@ -11,7 +11,7 @@
 #include <cstring>
 
 #include "c_mock_uploader.h"
-#include "compression.h"
+#include "compression/compression.h"
 #include "crypto/hash.h"
 #include "ingestion/item.h"
 #include "ingestion/item_mem.h"
